Se reescribió la clase usuario y se agregaron los atributos para guardar favoritos
